chore: fix instances of undefined left-shifting
Uses casts to prevent left shifts of negative numbers

Although casting from unsigned to signed is implementation defined, it
typically follows the usual behavior of 2's complement as is assumed
in this commit. GCC guarentees this behavior.

Relevant documentation: https://gcc.gnu.org/onlinedocs/gcc/Integers-implementation.html#Integers-implementation

+++ include/nds/arm9/math.h
@@ -71,7 +71,7 @@ extern "C" {
 ///     20.12 denominator.
 static inline void divf32_asynch(int32_t num, int32_t den)
 {
-    REG_DIV_NUMER = ((int64_t)num) << 12;
+    REG_DIV_NUMER = (int64_t)((uint64_t)(int64_t)num << 12);
     REG_DIV_DENOM_L = den;
 
     if ((REG_DIVCNT & DIV_MODE_MASK) != DIV_64_32)
